Introduce GPU resource cache.
This adds an incremental LRU cache that can be used to store values
that vertex shaders need to access.

The intent is to replace the use of the GPU store type, which
currently uploads the entire buffer each frame, and has a lot
of limitations (such as requiring allocation up front).

The new interface supports building GPU data on demand, as required,
which will improve CPU a lot for large pages.

For now, I have ported borders and rectangles to use the new type.
The patch is already large enough - this will allow us to port
each primitive type to use the GPU cache incrementally. The current
code also doesn't actually take advantage of building the data on
demand yet - they are just copied when needed.

Follow up work includes:
 * Porting more types to use the GPU cache.
 * Removing the old gpu store interfaces.
 * Profiling and optimizing various GL incremental update APIs.
 * Supporting resizing of the GPU cache backing store.