Return the clip id from push_clip_node.
In cases where the caller doesn't pass in an explicit clip id, it is
still useful for the caller to get back the generated clip id. This is
because the caller may later need to call push_clip_and_scroll_info with
the clip id that was generated. If this function doesn't return the clip
id the caller would basically have to inline this function and define
the clip manually, which seems much less convenient.