Move PrimitiveGeometry to be stored inline in the GPU cache.
The local rect and local clip rect are stored as a common header
to each primitive in the GPU cache. This removes the need for
providing the global primitive index in the primitive instances.

Also add a method for invalidating an entry in the GPU cache, which
is used for updating scroll bar primitives.

Add a few extra free lists so that we avoid using entire cache
rows for items > 8 blocks.