Start splitting apart ScrollFrames and Clips
ScrollFrames and clips are currently represented by one node type in
the ClipScrollTree, which means that many nodes are doing more work
than they need to do. This makes even less sense in a world where all
clips are added to the ClipScrollTree and handled the same way in the
API and internally. This change is the first step to splitting apart
these two types of items.

In order to minimize the changes, existing clips/ScrollFrame are
pushed as two nodes. This might almost double the depth of the tree,
but a later patch should expose API for pushing a clip without a
ScrollFrame attached.