Modify the blob sample app to use the boilerplate code
This also changes the background color of the blob app because I didn't
think it was worth adding another optional argument to main_wrapper just
for this, as the exact bgcolor is not particularly relevant to the blob
sample app.

Also, pressing the 'p' key will now toggle profiler display in all
sample apps that use the boilerplate, since it seems like a handy
feature.