Remove resolve pass, port resource rects to GPU cache.
Each item in the texture cache now has a GPU cache handle. The resource
cache tracks which items have been requested this frame, and ensures
the GPU cache is updated for each of these items.

The GPU cache address for each item is then retrieved during batching,
which means that the primitive resolve pass can be completely removed.

Items that don't exist in the texture cache (certain external image
types and webgl textures) push per-frame blocks to the GPU cache
that are (optionally) deferred and resolved by the external callback
on the render thread.

This simplifies the frame generation code, and is also a significant
optimization to the amount of CPU time used building each frame.