Optimize CPU time spent in building glyph instances.
Previously, the CPU would query the dimensions of each glyph, and
then apply the dimensions offset from the rasterized glyph to the
instance data. This results in a lot of hash lookups to query the
dimensions (and also some additional math on the CPU).

Instead, allow storing "user data" with each texture cache item. This
allows the glyph offsets to be applied in the vertex shader instead
of on the CPU.

Because this is such a hot code path, it provides significant CPU time
wins - often up to 20% of the CPU time on real pages.