Make push_prebuilt_display_list a memcpy
Make push_prebuilt_display_list a memcpy by adding internal support
for nested display lists. We extend the existing replacements code to
replace references to the root scrolling node id with the id of the
node that surrounds the built display list. This a little complicated
because we maintain support for multiple levels of nesting. This can be
removed if we decide that it isn't necessary.

We also rename the method to push_nested_display_list to match the
internal naming.

Fixes #1378.