[RFC] Add a rawtest framework
This adds a rawtesting framework that lets tests do anything
that the webrender api allows. This allows writing tests
that do more sophisticated things then the yaml can express.

The framework includes a single test that tests that when
we draw a blob image a second time we get the same results.