Remove the per-item complex clipping and masking from the API
This is the first step to getting all masking operations into the
ClipScrollTree. We don't completely this feature, as it's still sued
for the extra clips. These will be moved into the ClipScrollTree in a
later change. This change allows us to fully eliminate the
DisplayListBuilder::push_clip_region API and push its functionality
into the define_clip API.

Most of the changes here are:
  1. Changes to the examples, wrench, and reftests to match the new API.
  2. Adapting the "split rectangle clipped by rounded rect"
     optimization to use the ClipScrollTree. This is unfortunately a
     bit more complicated now, but now includes any rounded rect clip
     instead of per-item ones.