Add a define_scroll_frame API
This makes it explicit whether or not you are creating a ScrollFrame
when using the API. Before this was based on the content size of the
clip, which was awkward and could lead to weird situations where you
created a ScrollFrame when you didn't mean to or the reverse.

Fixes #1335.