Make clip coordinates relative to stacking context
Clip coordinates previously used a weird coordinate space relative to
the origin of the display item. This was different than every other
coordinate specified in the display list, so we smooth over this
wrinkle. To do this we need to copy out the complex clips from the
auxiliary list and adjust them so that they are relative to the scroll
clip node.

Fixes #1090.
Fixes #1408.