Add a way to remove all resources from the resource cache that have a certain namespace.
This is a tad ugly but it makes it a lot easier in firefox when a content-process crashes, to clean up every resources that were associated to it. The alternative is to maintain a map of all of the image and font keys in use in gecko-land.

Using IdNamespace more aggressively