Tweak condition for warning
Fixes #1424

It should be fine to call finalize() on a non-scroll node, as long as
the scroll offset passed in zero (i.e. that we're not actually trying to
scroll it). It would *not* be fine to call set_scroll_origin on a
non-scroll node, but that function emits a warning for that scenario
already.