Only convert scroll id to nested if pipeline id is match.
If nested display list is used and this nested display list contains
iframe. In current implementation, we'll convert the root scroll node
inside the iframe to the nested display list's root scroll node. Thus,
the primitive position inside the iframe is wrong. So I add a simple
check to avoid this.