Add interim / basic support for text decorations in text shadows.
This commit adds the basic code structure for including a text
decoration line item in a text shadow. In the near future,
we'll be expanding this structure to support a line decoration
display item, which supports all CSS provided text decorations.

However, in the interim, Servo relies on solid rectangles being
part of a text-shadow context for some reftests. To allow a WR
update in Servo, this commit supports rendering solid rectangles
as part of a text-shadow context. This hack only works in the case
where the text shadow has a zero blur radius, which is enough
for the Servo reftests and will be removed very soon!