Improve wrench display list "replay" and DL/IPC timing.
Previously, wrench held on to a single DLBuilder object, and copied
this and finalize'ed it each time a DL was replayed. This has several
issues, and messes up the timings of the DL profiling.

Instead, change wrench to build the DL once, and store the built
display lists. This mostly involves plumbing the DL builder through
all the yaml reader functions.

Also, modify the DL timing code to handle the situation where a DL
is built once and sent multiple times for profiling purposes. Specifically,
record an extra timestamp just before starting the send operation and
accumulate the total time as a sum of each component.