Refactor freelist and fix texture updates.
Refactor the freelist implementation to be simpler. Specifically,
the handles returned from the freelist are strong handles, that
cannot be cloned or copied. This is prep work for some upcoming
refactoring of the texture cache allocator and implementation.

While doing this, I found some bugs in the texture update APIs
that are now fixed. There were two issues:
 * When reusing an existing texture ID, the texture cache info
   was not being updated in the item cache. This meant the old
   dimensions were being used.
 * When a standalone texture was updated, it wasn't getting
   initialized in the code path where the dimensions are different.

It's a bit difficult to reftest this right now (until we get
raw tests), so I've added an example that checks both texture
update code paths. Neither of those code paths were working
correctly before this change.