Refactor the boilerplate code and update examples.
Instead of free functions, there is now an Example trait that
each example must implement. This allows examples to store
state without resorting to globals, mutexes and such things.

The trait has two methods:

 * render - the example must create a display list and resource
            transaction update matching the current application state.

 * on_event - called when a glutin event is received. This can return
              true if the example wants render() to be called again
              after the event is handled.

Previously, it was not possible to modify or rebuild the display list
after the initial display list was created. Now, examples can choose
to rebuild a display list as required.