Introduce texture_allocator.rs, and use it for render target allocation.
I'm working on a new allocator for the texture cache, that is better
suited to a dynamic glyph / image cache.

However, the new allocator is not suitable for the needs of the
render task tree, which currently shares the same allocator.

This patch copies the TexturePage allocator implementation from
the texture cache, into texture_allocator.rs, renames it to
guillotine allocator, and switches the render task tree to use it.

It is an exact copy of the code, but with unnecessary features
removed (e.g. all the freeing and coalescing support).

This means we have a bit of code duplication temporarily, but
allows changing the texture cache allocator without affecting the
allocator used by the render task tree.

In the future, we will likely want to tune / modify the allocator
implementation for render tasks as well.