Add external image support for both 2D and 2D_ARRAY samplers.
This fixes an oversight on my part when adding the new texture
cache implementation. The webgl / offscreen rendering code has
not been updated to create array textures.

Add explicit external image support for both 2D and 2D_ARRAY
samplers. In the future, we may deprecate / remove the 2D
image buffer kind, once we have switched everything over to
use array textures, but this is a good interim fix for now.