Switch clip code and shaders to use typed GpuCacheAddress directly.
This saves a bit of work converting address types in both CPU and
vertex shader code.

It also makes the vertex types use the correct type for cache
addresses, instead of casting to an i32.

The intent is to incrementally move all vertex formats over to
using the 2D style addresses directly, and then remove the
duplicated code from prim_shared.glsl which allows fetching via
an i32.