Add support for primitive shaders referencing the render target caches.
This switches ps_cache_image over to use SourceTexture::CacheRGBA8.

The idea here is to enable primitives added during batch creation
to select a render cache target as a source texture. This is possible
now since the texture cache uses array textures, and therefore shares
the same sampler target as the render target cache.

The goal is to begin unifying some of the shaders - for example,
the ps_cache_image shader can be replaced by the normal ps_image
shader, which will reduce the number of shaders and shader switches.