Add support for shader includes.
If the first line of a shader is //import followed by a comma
separated list of glsl filenames, prepend those as includes
for the shader. The idea here is that we'll start splitting
up prim_shared.glsl etc into smaller blocks that are faster
to parse and compile, and just generally more manageable.

This also supports having the VS and FS in the same glsl file. This
commit ports the debug_font and debug_color shaders to this style.

If we go ahead with this format, we'll incrementally port other
shaders to use this style, which requires only one file per shader.