Add a basic display list treeview to the debugger.
This allows viewing the input set of display lists and documents
for a given WR instance.

The treeview itself is very basic, and the information provided
per item is minimal. Consider this a proof of concept which can
be expanded upon to provide more information.

Importantly, it demonstrates how to provide information to the
debugger websocket that is only available in the render backend
(the previous debug items were all available in the renderer
instance).