Introduce shared A8 target that is available to all passes.
Many tasks (such as box shadows) are simple tasks that don't
have dependencies on other tasks in the tree. This change
ensures that those tasks are always drawn in the first render pass.

A new sampler is introduced that is always bound to the shared
target as an input for all subsequent passes, for tasks that
opt in.

This is a performance win where a cacheable task would previously
be rendered in multiple passes. For example, a box shadow that is
used by pass 1 and 2 would previously be drawn twice, but with this
change the box shadow is drawn only once (if the box shadow cache
key is the same).

This is a performance win by itself, but is also some prep work
for some ideas I want to prototype related to how we manage
clip masks as part of the clip stack.