Fix box shadow regression.
The box shadow shader vertex format was incorrectly being set
to primitive instances instead of the new vertex format.

On some implementations, this happened to work correctly. On
mac, the default vertex attribute binding meant that this bug
showed up there.

This change removes the vertex format enum, which is redundant
with the vertex array kind enum, and fixes the box shadow shader
to bind to the correct vertex format.

As a follow up, I'll make the shader compilation code assert
if a vertex attribute binding fails.