Move remaining mask cache code into ClipSources, and simplify it.
We no longer need to copy the border sources etc into the mask
cache structures - they can be referenced directly from the clip sources.

We also use a separate GPU cache handle for each clip now, so the
ClipAddressRange code is no longer needed. This is possible since
we use the GpuCache now, which doesn't need to allocate space
during initialization.

This also simplifies the code in the clip batcher that adds
clip instances to the render targets.

Add support for multiple image masks (a by-product of removing
the init step of the mask cache code).