Introduce frame output support.
This adds a new feature to WR, that allows copying the output
of a given pipeline ID to an external texture. This is a feature
request from the webvr prototyping team.

The included example demonstrates how to use the API to copy the
output of an iframe to an external texture, which is then copied
onto the screen via custom GL commands.