Remove AlphaBatchKeyFlags, and some batching prep / tidying.
AlphaBatchKeyFlags previously stored flags to specify whether
the batch needs clipping and/or transform shaders.

There are a few problems with this:
 - Rectangle is the *only* primitive that cares about the clip
   flag. All other primitives apply clipping unconditionally on
   their alpha versions. Having the clip flag in the key that
   is shared just breaks batches for no reason. Instead, move
   the clip key to be in the Rectangle specific batch kind.
 - Several of the batch kinds (e.g. blend, composite) don't care
   about the transform type of the primitive. Considering this
   breaks batches for no reason (although rarely).

The fixes above do help, but this is mostly prep work for the
refactoring and optimizing I'm planning to do to unify several
of the shaders. The unified shaders will make different batching
decisions (e.g. just the presence of a transform won't imply
alpha pass - for internal segments of the primitive).