Initialize render targets to the required size, saving memory and clear time.
We have the information available that tells us the worst case
*actual* size requirement for each render pass array texture. Keep
track of that and only create targets of that size.

This is a quite significant GPU memory saving, and also a large
performance win on *some* GPU architectures (seems to be a good
benefit on the Intel GPUs I've tested with).