Introduce ClipChain
Clip chains are a linked list of ClipWorkItems that are used to avoid
having to recalculate the stack of clips necessary to render items when
switching between different ClipScrollNodes. The ClipChains are
calculated once on a per node basis during update_transform.

In order to make this happen we need to start tracking when
ClipScrollNodes share axis-aligned coordinate spaces. We can use these
ids to render rectangular clips into masks when necessary.