Support different radii and elliptical radii in box shadows.
Now that box shadows are rendered via a normal mask + blur,
we can just pass through radii for each corner to handle
this case. Right now, the brush mask assumes it will handle
the entire rect. The next step is to support segmenting
parts of the mask generation, which will greatly help with
the performance of the new box shadow implementation.

Fixes #1430.