Optimize box shadows - only apply blurs where they can affect the shadow.
Add support for regions to blur render tasks. These allow render tasks
to specifiy one or more regions where the blur should be applied. If
no blur region is supplied, the entire render task is blurred.

Also start expanding the Picture struct, with methods to prepare
it for rendering etc.

Finally, add the basic infrastructure for render tasks to specify
what color they should be cleared to. This allows tasks to specify
a non-default color, if required. This is useful for inset shadows
where we will avoid blurring the inside of the shadow rect.

For outset box shadows with a large blur radius and/or border radius
this patch makes them ~3-4x faster than previously. A follow up will
be to add the same blur regions for inset box shadows.