Switch ps_cache_image to be a brush shader.
This starts to expand on how the brush shader works, and
converts one of the simplest primitive shaders over to use it.

The fundamental difference between a primitive and brush shader is:

Primitive shaders have two versions: transform and axis-aligned.

Brush shaders have an opaque and alpha version. In the future,
the opaque brush shader can run on a transformed primitive to
handle the inner part of the primitive which doesn't need AA.