Use primitive clip masks only, instead of sharing clip chain masks.
We're not getting any noticeable benefit from trying to share clip
masks. In fact, we are often seeing very significant slowdowns
since we over-estimate the size needed for a shared clip mask.

This change drops the GPU time in #1817 from 40ms to 7ms on
my machine - this is an extreme case, but it's faster in every
site that I tested on locally.

This is also a good simplification for upcoming changes, where
we will be more aggressively removing clip masks when they
are redundant, and drawing them in small segments where possible.