Use constant color blend for subpx text without clips.
The current subpx rendering mode that supports clip masks is a
two pass technique. With the current batcher implementation,
this can result in a lot of batches on sites where text
bounding rects are a bit loose, and are detected as potentially
overlapping. This most often occurs on sites that use a lot
of small text.

As an interim solution, only use the clip mask supporting method
when a clip mask (or translucent text) is active. Otherwise, use
the old color blending method.

In the near(ish) future, we'll start using the dual source blending
extension for subpixel rendering on hardware that supports it. This
extension gives the best of both worlds - supporting clip masks and
being a single pass technique. It's supported on the vast majority
of our target hardware (perhaps even all of it).

On hardware that doesn't support dual source blending, using the
modes here which invoke a different technique depending on need
for alpha masking is a very reasonable tradeoff.

Additionally, future improvements to batching will improve this
fallback solution by being a bit smarter about clumping together
non-overlapping batches.

Fixes #1984.