Fix some issues with scrollbars
Scrollbars would scroll horizontally instead of sticking to the edge of
the frame and were also not visible at all on iframes. This change fixes
those issues by adding them to the containing reference frame. As part
of this change, iframe background colors are also attached to the
reference frame and only painted over the frame rect.