Remove ClipInfo
Since Clip nodes always have clips in their ClipSources, the is_masking
member isn't necessary in ClipInfo. Since only one member remains,
we just store the ClipSourcesHandle and remove ClipInfo entirely. We
also rename ClipSources::is_masking() to ClipSources::has_clips()
because it is possible that a ClipSources may not be rendered to a mask
if the clip is rectangular. This can make the previous name of the
method misleading.