Make all render task data structs contain a common base field.
Some render tasks can have a source task of different kinds. The
most common example is that the source task to a blur task can
be any of (a) a picture (b) an alpha task (c) a downscaling task.

To ensure this works correctly, we separate render task data into
a common base, followed by a small amount of task-specific data.

The blur tasks now only fetch the common task data, which is all
that they require.

This fixes a bug (that I don't think is actually occurring in any
tests), and is also prep work for the next step. Now that everything
is using Picture internally, we can remove the AlphaBatchTask
completely. This will allow us to start converting more primitives
over to use a smaller subset of brush shaders, which will alow us
to start doing the primitive segmentation and also get better
batching results.