Remove AlphaBatchTask, and introduce start of raster mode.
Most of the difficult work related to this was done in the
large picture PR that landed previously.

This removes the remnants of AlphaBatch tasks, and makes them
all run through Picture render tasks.

This also introduces the first part of the picture and shader
side support for selecting whether primitives on a given picture
are to be rasterized in screen space or a local space.