Remove manual interpolation of local space in transform shaders.
Instead, for vertex transform shaders we now write out:
 (world.xy, z * world.w, world.w)

This gives us constant Z across the polygon for draw ordering.

It also means we can use the hardware interpolators instead of
manually dividing by Z in the fragment shader for local_pos.

This is mostly preparation work for switching shaders over to
be brush primitives, but has the happy side effect of:

 * Fixing a small number of Gecko + Servo reftests.
 * Simpler and faster code.