Update README to reflect new Cargo patch functionality
Cargo now has the `patch` functionality to do local replacements. This
is more comfortable to use both because you don't need to specify the
version and because the `replace` section doesn't work as well as it
used it. The patch syntax is a bit funky though, so we should probably
have the README reference it directly.