Invert ClipScrollNode transformation in the vertex shader
This allows us to offload fairly expensive CPU work to the vertex
shader and only when get_layer_pos is called. Instead, we are able
to only check the determinant and then only for reference frames. This
seems like it would only be a problem if many clips are sharing the
same ClipScrollNode, which is an unusual situation.

In this change, we also stop doing unnecessary work on ClipScrollNodes
once they have an uninvertible parent or we know one of the parents is
clipped out entirely.